拆分main.py中测试节点的xray相关代码到新增的xray.py模块中
